feat: implement allowedMaxCost setting
Introduces the `allowedMaxCost` setting to control the maximum cost allowed for auto-approved requests.

This change includes:
- Adding `allowedMaxCost` to `GlobalSettingsSchema`, `ExtensionState`, and `WebviewMessage` types.
- Updating `Task.ts` to track `consecutiveAutoApprovedCost`.
- Modifying `ClineProvider.ts` to handle `allowedMaxCost` in state management.
- Updating `webviewMessageHandler.ts` to process `allowedMaxCost` messages.
- Creating new UI components `MaxCostInput.tsx` and `MaxLimitInputs.tsx` for the setting.
- Enhancing `FormattedTextField.tsx` with `unlimitedDecimalFormatter` for cost input.
- Updating `AutoApproveMenu.tsx` and `SettingsView.tsx` to integrate the new cost limit input.
- Adjusting `AutoApprovedRequestLimitWarning.tsx` to differentiate between request and cost limits.
- Adding `kilocode_change` markers where necessary for upstream compatibility.
