+++ packages/types/src/global-settings.ts
@@ -69,7 +69,7 @@ export const globalSettingsSchema = z.object({
 	commandTimeoutAllowlist: z.array(z.string()).optional(),
 	preventCompletionWithOpenTodos: z.boolean().optional(),
 	allowedMaxRequests: z.number().nullish(),
-	allowedMaxCost: z.number().nullish(),
+	allowedMaxCost: z.number().nullish(), // kilocode_change
 	autoCondenseContext: z.boolean().optional(),
 	autoCondenseContextPercent: z.number().optional(),
 	maxConcurrentFileReads: z.number().optional(),
