+++ src/shared/ExtensionMessage.ts
@@ -249,7 +249,7 @@ export type ExtensionState = Pick<
 	| "allowedCommands"
 	| "deniedCommands"
 	| "allowedMaxRequests"
-	| "allowedMaxCost"
+	| "allowedMaxCost" // kilocode_change
 	| "browserToolEnabled"
 	| "browserViewportSize"
 	| "showAutoApproveMenu" // kilocode_change
