+++ webview-ui/src/components/common/DecoratedVSCodeTextField.tsx
@@ -1,3 +1,4 @@
+// kilcode_change - new file
 import { cn } from "@/lib/utils"
 import { VSCodeTextField } from "@vscode/webview-ui-toolkit/react"
 import { forwardRef, useCallback, useRef, ReactNode, ComponentRef, ComponentProps } from "react"
