+++ webview-ui/src/components/settings/__tests__/MaxCostInput.spec.tsx
@@ -1,3 +1,4 @@
+// kilocode_change - new file
 import { render, screen, fireEvent } from "@testing-library/react"
 import { vi } from "vitest"
 import { MaxCostInput } from "../MaxCostInput"
